arch/xtensa/esp32: use common wireless source
Remove ESP32 specific WiFi files and edit build system to use common layer.
