+++ arch/xtensa/src/common/espressif/esp_wifi_utils.c
@@ -34,6 +34,9 @@
 #include <nuttx/kmalloc.h>
 #include <nuttx/wireless/wireless.h>
 
+#ifdef CONFIG_ARCH_CHIP_ESP32
+#include "esp32_wifi_adapter.h"
+#endif
 #ifdef CONFIG_ARCH_CHIP_ESP32S2
 #include "esp32s2_wifi_adapter.h"
 #endif
