+++ arch/xtensa/src/common/espressif/esp_wlan.h
@@ -29,6 +29,9 @@
 
 #include <nuttx/config.h>
 
+#ifdef CONFIG_ARCH_CHIP_ESP32
+#  include "esp32_wifi_adapter.h"
+#endif
 #ifdef CONFIG_ARCH_CHIP_ESP32S2
 #  include "esp32s2_wifi_adapter.h"
 #endif
