+++ arch/xtensa/src/esp32/Kconfig
@@ -734,33 +734,6 @@ config ESP32_UART2
 	select UART2_SERIALDRIVER
 	select ARCH_HAVE_SERIAL_TERMIOS
 
-config ESP32_WIRELESS
-	bool
-	default n
-	select NET
-	select ARCH_PHY_INTERRUPT
-	select ESP32_RNG
-	select ESP32_RT_TIMER
-	select ESP32_TIMER0
-	---help---
-		Enable Wireless support
-
-config ESP32_WIFI
-	bool "Wi-Fi"
-	default n
-	select ESP32_WIRELESS
-	---help---
-		Enable Wi-Fi support
-
-config ESP32_BLE
-	bool "BLE"
-	default n
-	select ESP32_WIRELESS
-	select BLUETOOTH_TXCMD_PINNED_TO_CORE if SMP
-	select BLUETOOTH_CNTRL_HOST_FLOW_DISABLE
-	---help---
-		Enable BLE support
-
 config ESP32_I2C0
 	bool "I2C 0"
 	default n
@@ -792,13 +765,6 @@ config ESP32_PID
 
 endmenu # ESP32 Peripheral Selection
 
-config ESP32_WIFI_BT_COEXIST
-	bool "Wi-Fi and BT coexist"
-	default y if ESP32_WIFI && ESP32_BLE
-	default n
-	depends on ESP32_WIFI && ESP32_BLE
-	select ESP32_WIFI_STA_DISCONNECT_PM
-
 menu "Interrupt Configuration"
 
 config ESP32_IRAM_ISR_DEBUG
@@ -816,8 +782,8 @@ menu "Memory Configuration"
 
 config ESP32_BT_RESERVE_DRAM
 	int "Reserved BT DRAM"
-	default 0 if !ESP32_BLE
-	default 65536 if ESP32_BLE
+	default 0 if !ESPRESSIF_BLE
+	default 65536 if ESPRESSIF_BLE
 
 config ESP32_TRACEMEM_RESERVE_DRAM
 	int "Reserved trace memory DRAM"
@@ -2074,142 +2040,10 @@ config ESP32_ETH_PHY_ADDR
 
 endmenu # ESP32_EMAC
 
-menu "Wi-Fi Configuration"
-	depends on ESP32_WIFI
-
-choice
-	prompt "ESP32 Wi-Fi mode"
-	default ESP32_WIFI_STATION
-
-config ESP32_WIFI_STATION
-	bool "Station mode"
-
-config ESP32_WIFI_SOFTAP
-	bool "SoftAP mode"
-
-config ESP32_WIFI_STATION_SOFTAP
-	bool "Station + SoftAP"
-
-endchoice # ESP32 Wi-Fi mode
-
-config ESP32_WIFI_STATIC_RXBUF_NUM
-	int "Wi-Fi static RX buffer number"
-	default 10
-
-config ESP32_WIFI_DYNAMIC_RXBUF_NUM
-	int "Wi-Fi dynamic RX buffer number"
-	default 32
-
-config ESP32_WIFI_DYNAMIC_TXBUF_NUM
-	int "Wi-Fi dynamic TX buffer number"
-	default 32
-
-config ESP32_WIFI_TX_AMPDU
-	bool "Wi-Fi TX AMPDU"
-	default y
-
-config ESP32_WIFI_RX_AMPDU
-	bool "Wi-Fi RX AMPDU"
-	default y
-
-config ESP32_WIFI_RXBA_AMPDU_WZ
-	int "Wi-Fi RX BA AMPDU windown size"
-	default 6
-
-config ESP32_WLAN_PKTBUF_NUM
-	int "WLAN netcard packet buffer number per netcard"
-	default 16
-
-config ESP32_WIFI_CONNECT_TIMEOUT
-	int "Connect timeout by second"
-	default 10
-	---help---
-		Max waiting time of connecting to AP.
-
-config ESP32_WIFI_SCAN_RESULT_SIZE
-	int "Scan result buffer"
-	default 4096
-	---help---
-		Maximum scan result buffer size.
-
-config ESP32_WIFI_STA_DISCONNECT_PM
-	bool "Power Management for station when disconnected"
-	default n
-	---help---
-		Select this option to enable power management for station when disconnected.
-		Chip will do modem-sleep when RF module is not in use anymore.
-
-config EXAMPLE_WIFI_LISTEN_INTERVAL
-	int "Wi-Fi listen interval"
-	default 3
-	---help---
-		Interval for station to listen to beacon from AP. The unit of listen interval is one beacon interval.
-		For example, if beacon interval is 100 ms and listen interval is 3, the interval for station to listen
-		to beacon is 300 ms.
-
-choice EXAMPLE_POWER_SAVE_MODE
-	prompt "Power save mode"
-	default EXAMPLE_POWER_SAVE_NONE
-	---help---
-		Power save mode for the esp32 to use. Modem sleep mode includes minimum and maximum power save modes.
-		In minimum power save mode, station wakes up every DTIM to receive beacon. Broadcast data will not be
-		lost because it is transmitted after DTIM. However, it can not save much more power if DTIM is short
-		for DTIM is determined by AP.
-		In maximum power save mode, station wakes up every listen interval to receive beacon. Broadcast data
-		may be lost because station may be in sleep state at DTIM time. If listen interval is longer, more power
-		is saved but broadcast data is more easy to lose.
-
-config EXAMPLE_POWER_SAVE_NONE
-	bool "none"
-
-config EXAMPLE_POWER_SAVE_MIN_MODEM
-	bool "minimum modem"
-
-config EXAMPLE_POWER_SAVE_MAX_MODEM
-	bool "maximum modem"
-
-endchoice # EXAMPLE_POWER_SAVE_MODE
-
-config ESP32_WIFI_WLAN_BUFFER_OPTIMIZATION
-	bool "Enable optimization of WLAN driver buffer"
-	default n
-	---help---
-		Enable optimization of WLAN memory
-
-endmenu # ESP32_WIFI
-
-menu "BLE Configuration"
-	depends on ESP32_BLE
-
-config ESP32_BLE_PKTBUF_NUM
-	int "BLE netcard packet buffer number per netcard"
-	default 16
-
-config ESP32_BLE_TASK_STACK_SIZE
-	int "Controller task stack size"
-	default 4096
-
-config ESP32_BLE_TASK_PRIORITY
-	int "Controller task priority"
-	default 110
-
-config ESP32_BLE_INTERRUPT_SAVE_STATUS
-	int "Number of interrupt save status"
-	default 3
-	---help---
-		Number of interrupt save status variables to keep track. Increase it if any related bug is found.
-
-config ESP32_BLE_MAX_CONN
-	int "Maximum BLE simultaneous connections"
-	range 1 3
-	default 1
-
-endmenu # BLE Configuration
-
 choice ESP32_UNIVERSAL_MAC_ADDRESSES
 	bool "Number of universally administered (by IEEE) MAC addresses"
 	default ESP32_UNIVERSAL_MAC_ADDRESSES_FOUR
-	depends on ESP32_WIFI || ESP32_BLE
+	depends on ESPRESSIF_WIFI || ESPRESSIF_BLE
 	---help---
 		Configure the number of universally administered (by IEEE) MAC addresses.
 		During initialization, MAC addresses for each network interface are generated or derived from a
