+++ arch/xtensa/src/esp32/esp32_wifi_adapter.c
@@ -58,19 +58,19 @@
 #include "hardware/esp32_dport.h"
 #include "hardware/esp32_emac.h"
 #include "esp32_irq.h"
-#include "esp32_wireless.h"
+#include "espressif/esp_wireless.h"
 #include "esp32_wifi_adapter.h"
 #include "esp32_rt_timer.h"
-#include "esp32_wifi_utils.h"
-#include "esp32_wlan.h"
+#include "espressif/esp_wifi_utils.h"
+#include "espressif/esp_wlan.h"
 
 #ifdef CONFIG_PM
 #  include "esp32_pm.h"
 #endif
 
-#ifdef CONFIG_ESP32_BLE
+#ifdef CONFIG_ESPRESSIF_BLE
 #  include "esp32_ble_adapter.h"
-#  ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#  ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
 #    include "private/esp_coexist_internal.h"
 #  endif
 #endif
@@ -92,7 +92,7 @@
 
 #define PHY_RF_MASK   ((1 << PHY_BT_MODULE) | (1 << PHY_WIFI_MODULE))
 
-#define WIFI_CONNECT_TIMEOUT  CONFIG_ESP32_WIFI_CONNECT_TIMEOUT
+#define WIFI_CONNECT_TIMEOUT  CONFIG_ESPRESSIF_WIFI_CONNECT_TIMEOUT
 
 #define TIMER_INITIALIZED_VAL (0x5aa5a55a)
 
@@ -111,11 +111,11 @@
 
 /* CONFIG_POWER_SAVE_MODEM */
 
-#if defined(CONFIG_ESP32_POWER_SAVE_MIN_MODEM)
+#if defined(CONFIG_ESPRESSIF_POWER_SAVE_MIN_MODEM)
 #  define DEFAULT_PS_MODE WIFI_PS_MIN_MODEM
-#elif defined(CONFIG_ESP32_POWER_SAVE_MAX_MODEM)
+#elif defined(CONFIG_ESPRESSIF_POWER_SAVE_MAX_MODEM)
 #  define DEFAULT_PS_MODE WIFI_PS_MAX_MODEM
-#elif defined(CONFIG_ESP32_POWER_SAVE_NONE)
+#elif defined(CONFIG_ESPRESSIF_POWER_SAVE_NONE)
 #  define DEFAULT_PS_MODE WIFI_PS_NONE
 #else
 #  define DEFAULT_PS_MODE WIFI_PS_NONE
@@ -193,11 +193,11 @@ struct nvs_adpt
  * Private Function Prototypes
  ****************************************************************************/
 
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
 static int semphr_take_from_isr_wrapper(void *semphr, void *hptw);
 static int semphr_give_from_isr_wrapper(void *semphr, void *hptw);
 static int is_in_isr_wrapper(void);
-#endif /* CONFIG_ESP32_WIFI_BT_COEXIST */
+#endif /* CONFIG_ESPRESSIF_WIFI_BT_COEXIST */
 
 static bool wifi_env_is_chip(void);
 static void wifi_set_intr(int32_t cpu_no, uint32_t intr_source,
@@ -360,7 +360,7 @@ static mutex_t g_wifiexcl_lock = NXMUTEX_INITIALIZER;
 
 static int g_wifi_ref;
 
-#ifdef ESP32_WLAN_HAS_STA
+#ifdef ESPRESSIF_WLAN_HAS_STA
 
 /* If reconnect automatically */
 
@@ -382,9 +382,9 @@ static wifi_txdone_cb_t g_sta_txdone_cb;
 
 static wifi_config_t g_sta_wifi_cfg;
 
-#endif /* ESP32_WLAN_HAS_STA */
+#endif /* ESPRESSIF_WLAN_HAS_STA */
 
-#ifdef ESP32_WLAN_HAS_SOFTAP
+#ifdef ESPRESSIF_WLAN_HAS_SOFTAP
 
 /* If Wi-Fi SoftAP starts */
 
@@ -398,7 +398,7 @@ static wifi_txdone_cb_t g_softap_txdone_cb;
 
 static wifi_config_t g_softap_wifi_cfg;
 
-#endif /* ESP32_WLAN_HAS_SOFTAP */
+#endif /* ESPRESSIF_WLAN_HAS_SOFTAP */
 
 /* Device specific lock */
 
@@ -410,7 +410,7 @@ static spinlock_t g_lock;
 
 /* Wi-Fi and BT coexistence OS adapter data */
 
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
 coex_adapter_funcs_t g_coex_adapter_funcs =
 {
   ._version = COEX_ADAPTER_VERSION,
@@ -435,7 +435,7 @@ coex_adapter_funcs_t g_coex_adapter_funcs =
   ._timer_arm_us = esp_timer_arm_us,
   ._magic = COEX_ADAPTER_MAGIC,
 };
-#endif /* CONFIG_ESP32_WIFI_BT_COEXIST */
+#endif /* CONFIG_ESPRESSIF_WIFI_BT_COEXIST */
 
 /* Wi-Fi OS adapter data */
 
@@ -499,7 +499,7 @@ wifi_osi_funcs_t g_wifi_osi_funcs =
   ._phy_enable = esp_phy_enable_wrapper,
   ._phy_common_clock_enable = esp_phy_common_clock_enable,
   ._phy_common_clock_disable = esp_phy_common_clock_disable,
-  ._phy_update_country_info = esp32_phy_update_country_info,
+  ._phy_update_country_info = esp_phy_update_country_info,
   ._read_mac = esp_wifi_read_mac,
   ._timer_arm = esp_timer_arm,
   ._timer_disarm = esp_timer_disarm,
@@ -2043,7 +2043,7 @@ static int esp_event_id_map(int event_id)
         id = WIFI_ADPT_EVT_SCAN_DONE;
         break;
 
-#ifdef ESP32_WLAN_HAS_STA
+#ifdef ESPRESSIF_WLAN_HAS_STA
       case WIFI_EVENT_STA_START:
         id = WIFI_ADPT_EVT_STA_START;
         break;
@@ -2063,9 +2063,9 @@ static int esp_event_id_map(int event_id)
       case WIFI_EVENT_STA_STOP:
         id = WIFI_ADPT_EVT_STA_STOP;
         break;
-#endif /* ESP32_WLAN_HAS_STA */
+#endif /* ESPRESSIF_WLAN_HAS_STA */
 
-#ifdef ESP32_WLAN_HAS_SOFTAP
+#ifdef ESPRESSIF_WLAN_HAS_SOFTAP
       case WIFI_EVENT_AP_START:
         id = WIFI_ADPT_EVT_AP_START;
         break;
@@ -2081,7 +2081,7 @@ static int esp_event_id_map(int event_id)
       case WIFI_EVENT_AP_STADISCONNECTED:
         id = WIFI_ADPT_EVT_AP_STADISCONNECTED;
         break;
-#endif /* ESP32_WLAN_HAS_SOFTAP */
+#endif /* ESPRESSIF_WLAN_HAS_SOFTAP */
 
       default:
         return -1;
@@ -2122,7 +2122,7 @@ static void esp_evt_work_cb(void *arg)
           break;
         }
 
-      /* Some of the following logic (eg. esp32_wlan_sta_set_linkstatus)
+      /* Some of the following logic (eg. esp_wlan_sta_set_linkstatus)
        * can take net_lock(). To maintain the consistent locking order,
        * we take net_lock() here before taking esp_wifi_lock. Note that
        * net_lock() is a recursive lock.
@@ -2137,13 +2137,13 @@ static void esp_evt_work_cb(void *arg)
             esp_wifi_scan_event_parse();
             break;
 
-#ifdef ESP32_WLAN_HAS_STA
+#ifdef ESPRESSIF_WLAN_HAS_STA
           case WIFI_ADPT_EVT_STA_START:
             wlinfo("Wi-Fi sta start\n");
 
             g_sta_connected = false;
 
-#ifdef CONFIG_ESP32_BLE
+#ifdef CONFIG_ESPRESSIF_BLE
             if (esp32_bt_controller_get_status() !=
                 ESP_BT_CONTROLLER_STATUS_IDLE)
               {
@@ -2175,7 +2175,7 @@ static void esp_evt_work_cb(void *arg)
           case WIFI_ADPT_EVT_STA_CONNECT:
             wlinfo("Wi-Fi sta connect\n");
             g_sta_connected = true;
-            ret = esp32_wlan_sta_set_linkstatus(true);
+            ret = esp_wlan_sta_set_linkstatus(true);
             if (ret < 0)
               {
                 wlerr("ERROR: Failed to set Wi-Fi station link status\n");
@@ -2186,7 +2186,7 @@ static void esp_evt_work_cb(void *arg)
           case WIFI_ADPT_EVT_STA_DISCONNECT:
             wlinfo("Wi-Fi sta disconnect\n");
             g_sta_connected = false;
-            ret = esp32_wlan_sta_set_linkstatus(false);
+            ret = esp_wlan_sta_set_linkstatus(false);
             if (ret < 0)
               {
                 wlerr("ERROR: Failed to set Wi-Fi station link status\n");
@@ -2206,13 +2206,13 @@ static void esp_evt_work_cb(void *arg)
             wlinfo("Wi-Fi sta stop\n");
             g_sta_connected = false;
             break;
-#endif /* ESP32_WLAN_HAS_STA */
+#endif /* ESPRESSIF_WLAN_HAS_STA */
 
-#ifdef ESP32_WLAN_HAS_SOFTAP
+#ifdef ESPRESSIF_WLAN_HAS_SOFTAP
           case WIFI_ADPT_EVT_AP_START:
             wlinfo("INFO: Wi-Fi softap start\n");
 
-#ifdef CONFIG_ESP32_BLE
+#ifdef CONFIG_ESPRESSIF_BLE
             if (esp32_bt_controller_get_status() !=
                 ESP_BT_CONTROLLER_STATUS_IDLE)
               {
@@ -2252,7 +2252,7 @@ static void esp_evt_work_cb(void *arg)
           case WIFI_ADPT_EVT_AP_STADISCONNECTED:
             wlinfo("INFO: Wi-Fi station leave\n");
             break;
-#endif /* ESP32_WLAN_HAS_SOFTAP */
+#endif /* ESPRESSIF_WLAN_HAS_SOFTAP */
           default:
             break;
         }
@@ -2278,7 +2278,7 @@ static void esp_evt_work_cb(void *arg)
     }
 }
 
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
 
 /****************************************************************************
  * Name: semphr_take_from_isr_wrapper
@@ -2343,7 +2343,7 @@ static int IRAM_ATTR is_in_isr_wrapper(void)
   return up_interrupt_context();
 }
 
-#endif /* CONFIG_ESP32_WIFI_BT_COEXIST */
+#endif /* CONFIG_ESPRESSIF_WIFI_BT_COEXIST */
 
 /****************************************************************************
  * Name: wifi_env_is_chip
@@ -3619,7 +3619,7 @@ static void esp_wifi_delete_queue(void *queue)
 
 static int coex_init_wrapper(void)
 {
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
   return coex_init();
 #else
   return 0;
@@ -3642,7 +3642,7 @@ static int coex_init_wrapper(void)
 
 static void coex_deinit_wrapper(void)
 {
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
   coex_deinit();
 #endif
 }
@@ -3664,7 +3664,7 @@ static void coex_deinit_wrapper(void)
 
 static int coex_enable_wrapper(void)
 {
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
   return coex_enable();
 #else
   return 0;
@@ -3687,7 +3687,7 @@ static int coex_enable_wrapper(void)
 
 static void coex_disable_wrapper(void)
 {
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
   coex_disable();
 #endif
 }
@@ -3709,7 +3709,7 @@ static void coex_disable_wrapper(void)
 
 static IRAM_ATTR uint32_t coex_status_get_wrapper(void)
 {
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
   return coex_status_get();
 #else
   return 0;
@@ -3736,7 +3736,7 @@ static IRAM_ATTR uint32_t coex_status_get_wrapper(void)
 static int32_t coex_wifi_request_wrapper(uint32_t event, uint32_t latency,
                                          uint32_t duration)
 {
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
   return coex_wifi_request(event, latency, duration);
 #else
   return 0;
@@ -3760,7 +3760,7 @@ static int32_t coex_wifi_request_wrapper(uint32_t event, uint32_t latency,
 
 static IRAM_ATTR int32_t coex_wifi_release_wrapper(uint32_t event)
 {
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
   return coex_wifi_release(event);
 #else
   return 0;
@@ -3785,7 +3785,7 @@ static IRAM_ATTR int32_t coex_wifi_release_wrapper(uint32_t event)
 
 static int coex_wifi_channel_set_wrapper(uint8_t primary, uint8_t secondary)
 {
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
   return coex_wifi_channel_set(primary, secondary);
 #else
   return 0;
@@ -3811,7 +3811,7 @@ static int coex_wifi_channel_set_wrapper(uint8_t primary, uint8_t secondary)
 static int coex_event_duration_get_wrapper(uint32_t event,
                                            uint32_t *duration)
 {
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
   return coex_event_duration_get(event, duration);
 #else
   return 0;
@@ -3857,7 +3857,7 @@ static int coex_pti_get_wrapper(uint32_t event, uint8_t *pti)
 static void coex_schm_status_bit_clear_wrapper(uint32_t type,
                                                uint32_t status)
 {
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
   coex_schm_status_bit_clear(type, status);
 #endif
 }
@@ -3879,7 +3879,7 @@ static void coex_schm_status_bit_clear_wrapper(uint32_t type,
 
 static void coex_schm_status_bit_set_wrapper(uint32_t type, uint32_t status)
 {
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
   coex_schm_status_bit_set(type, status);
 #endif
 }
@@ -3901,7 +3901,7 @@ static void coex_schm_status_bit_set_wrapper(uint32_t type, uint32_t status)
 
 static IRAM_ATTR int coex_schm_interval_set_wrapper(uint32_t interval)
 {
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
   return coex_schm_interval_set(interval);
 #else
   return 0;
@@ -3924,7 +3924,7 @@ static IRAM_ATTR int coex_schm_interval_set_wrapper(uint32_t interval)
 
 static uint32_t coex_schm_interval_get_wrapper(void)
 {
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
   return coex_schm_interval_get();
 #else
   return 0;
@@ -3947,7 +3947,7 @@ static uint32_t coex_schm_interval_get_wrapper(void)
 
 static uint8_t coex_schm_curr_period_get_wrapper(void)
 {
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
   return coex_schm_curr_period_get();
 #else
   return 0;
@@ -3970,7 +3970,7 @@ static uint8_t coex_schm_curr_period_get_wrapper(void)
 
 static void *coex_schm_curr_phase_get_wrapper(void)
 {
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
   return coex_schm_curr_phase_get();
 #else
   return NULL;
@@ -3994,7 +3994,7 @@ static void *coex_schm_curr_phase_get_wrapper(void)
 
 static int coex_register_start_cb_wrapper(int (* cb)(void))
 {
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
   return coex_register_start_cb(cb);
 #else
   return 0;
@@ -4018,7 +4018,7 @@ static int coex_register_start_cb_wrapper(int (* cb)(void))
 
 static int coex_schm_process_restart_wrapper(void)
 {
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
   return coex_schm_process_restart();
 #else
   return 0;
@@ -4043,7 +4043,7 @@ static int coex_schm_process_restart_wrapper(void)
 
 static int coex_schm_register_cb_wrapper(int type, int(*cb)(int))
 {
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
   return coex_schm_register_callback(type, cb);
 #else
   return 0;
@@ -4145,7 +4145,7 @@ static unsigned long esp_random_ulong(void)
 static IRAM_ATTR void esp_wifi_tx_done_cb(uint8_t ifidx, uint8_t *data,
                                           uint16_t *len, bool txstatus)
 {
-#ifdef ESP32_WLAN_HAS_STA
+#ifdef ESPRESSIF_WLAN_HAS_STA
   if (ifidx == ESP_IF_WIFI_STA)
     {
       if (g_sta_txdone_cb)
@@ -4154,9 +4154,9 @@ static IRAM_ATTR void esp_wifi_tx_done_cb(uint8_t ifidx, uint8_t *data,
         }
     }
   else
-#endif /* ESP32_WLAN_HAS_STA */
+#endif /* ESPRESSIF_WLAN_HAS_STA */
 
-#ifdef ESP32_WLAN_HAS_SOFTAP
+#ifdef ESPRESSIF_WLAN_HAS_SOFTAP
   if (ifidx == ESP_IF_WIFI_AP)
     {
       if (g_softap_txdone_cb)
@@ -4165,13 +4165,13 @@ static IRAM_ATTR void esp_wifi_tx_done_cb(uint8_t ifidx, uint8_t *data,
         }
     }
   else
-#endif /* ESP32_WLAN_HAS_SOFTAP */
+#endif /* ESPRESSIF_WLAN_HAS_SOFTAP */
     {
       wlerr("ifidx=%d is error\n", ifidx);
     }
 }
 
-#ifdef ESP32_WLAN_HAS_STA
+#ifdef ESPRESSIF_WLAN_HAS_STA
 
 /****************************************************************************
  * Name: esp_wifi_auth_trans
@@ -4273,7 +4273,7 @@ static int esp_wifi_cipher_trans(uint32_t wifi_cipher)
   return cipher_mode;
 }
 
-#endif /* ESP32_WLAN_HAS_STA */
+#endif /* ESPRESSIF_WLAN_HAS_STA */
 
 /****************************************************************************
  * Name: esp_freq_to_channel
@@ -4609,28 +4609,28 @@ int esp_wifi_adapter_init(void)
 
   wifi_cfg.nvs_enable = 0;
 
-#ifdef CONFIG_ESP32_WIFI_TX_AMPDU
+#ifdef CONFIG_ESPRESSIF_WIFI_TX_AMPDU
   wifi_cfg.ampdu_tx_enable = 1;
 #else
   wifi_cfg.ampdu_tx_enable = 0;
 #endif
 
-#ifdef CONFIG_ESP32_WIFI_RX_AMPDU
+#ifdef CONFIG_ESPRESSIF_WIFI_RX_AMPDU
   wifi_cfg.ampdu_rx_enable = 1;
 #else
   wifi_cfg.ampdu_rx_enable = 0;
 #endif
 
-#ifdef CONFIG_ESP32_WIFI_STA_DISCONNECT_PM
+#ifdef CONFIG_ESPRESSIF_WIFI_STA_DISCONNECT_PM
   wifi_cfg.sta_disconnected_pm = true;
 #else
   wifi_cfg.sta_disconnected_pm = false;
 #endif
 
-  wifi_cfg.rx_ba_win          = CONFIG_ESP32_WIFI_RXBA_AMPDU_WZ;
-  wifi_cfg.static_rx_buf_num  = CONFIG_ESP32_WIFI_STATIC_RXBUF_NUM;
-  wifi_cfg.dynamic_rx_buf_num = CONFIG_ESP32_WIFI_DYNAMIC_RXBUF_NUM;
-  wifi_cfg.dynamic_tx_buf_num = CONFIG_ESP32_WIFI_DYNAMIC_TXBUF_NUM;
+  wifi_cfg.rx_ba_win          = CONFIG_ESPRESSIF_WIFI_RXBA_AMPDU_WZ;
+  wifi_cfg.static_rx_buf_num  = CONFIG_ESPRESSIF_WIFI_STATIC_RXBUF_NUM;
+  wifi_cfg.dynamic_rx_buf_num = CONFIG_ESPRESSIF_WIFI_DYNAMIC_RXBUF_NUM;
+  wifi_cfg.dynamic_tx_buf_num = CONFIG_ESPRESSIF_WIFI_DYNAMIC_TXBUF_NUM;
 
   ret = esp_wifi_init(&wifi_cfg);
   if (ret)
@@ -4668,7 +4668,7 @@ errout_init_wifi:
  * Station functions
  ****************************************************************************/
 
-#ifdef ESP32_WLAN_HAS_STA
+#ifdef ESPRESSIF_WLAN_HAS_STA
 
 /****************************************************************************
  * Name: esp_wifi_sta_start
@@ -4698,13 +4698,13 @@ int esp_wifi_sta_start(void)
       wlinfo("Failed to stop Wi-Fi ret=%d\n", ret);
     }
 
-#ifdef ESP32_WLAN_HAS_SOFTAP
+#ifdef ESPRESSIF_WLAN_HAS_SOFTAP
   if (g_softap_started)
     {
       mode = WIFI_MODE_APSTA;
     }
   else
-#endif /* ESP32_WLAN_HAS_SOFTAP */
+#endif /* ESPRESSIF_WLAN_HAS_SOFTAP */
     {
       mode = WIFI_MODE_STA;
     }
@@ -4763,7 +4763,7 @@ int esp_wifi_sta_stop(void)
 
   g_sta_started = false;
 
-#ifdef ESP32_WLAN_HAS_SOFTAP
+#ifdef ESPRESSIF_WLAN_HAS_SOFTAP
   if (g_softap_started)
     {
       ret = esp_wifi_set_mode(WIFI_MODE_AP);
@@ -4782,13 +4782,13 @@ int esp_wifi_sta_stop(void)
           goto errout;
         }
     }
-#endif /* ESP32_WLAN_HAS_SOFTAP */
+#endif /* ESPRESSIF_WLAN_HAS_SOFTAP */
 
   wlinfo("OK to stop Wi-Fi station\n");
 
-#ifdef ESP32_WLAN_HAS_SOFTAP
+#ifdef ESPRESSIF_WLAN_HAS_SOFTAP
 errout:
-#endif /* ESP32_WLAN_HAS_SOFTAP */
+#endif /* ESPRESSIF_WLAN_HAS_SOFTAP */
 
   esp_wifi_lock(false);
   return ret;
@@ -5663,7 +5663,7 @@ int esp_wifi_sta_bitrate(struct iwreq *iwr, bool set)
   return OK;
 }
 
-#endif /* ESP32_WLAN_HAS_STA */
+#endif /* ESPRESSIF_WLAN_HAS_STA */
 
 /****************************************************************************
  * Name: esp_wifi_sta_get_txpower
@@ -5858,7 +5858,7 @@ int esp_wifi_sta_country(struct iwreq *iwr, bool set)
   return OK;
 }
 
-#ifdef ESP32_WLAN_HAS_STA
+#ifdef ESPRESSIF_WLAN_HAS_STA
 
 /****************************************************************************
  * Name: esp_wifi_sta_rssi
@@ -5905,13 +5905,13 @@ int esp_wifi_sta_rssi(struct iwreq *iwr, bool set)
 
   return OK;
 }
-#endif /* ESP32_WLAN_HAS_STA */
+#endif /* ESPRESSIF_WLAN_HAS_STA */
 
 /****************************************************************************
  * SoftAP functions
  ****************************************************************************/
 
-#ifdef ESP32_WLAN_HAS_SOFTAP
+#ifdef ESPRESSIF_WLAN_HAS_SOFTAP
 
 /****************************************************************************
  * Name: esp_wifi_softap_start
@@ -5941,13 +5941,13 @@ int esp_wifi_softap_start(void)
       wlinfo("Failed to stop Wi-Fi ret=%d\n", ret);
     }
 
-#ifdef ESP32_WLAN_HAS_STA
+#ifdef ESPRESSIF_WLAN_HAS_STA
   if (g_sta_started)
     {
       mode = WIFI_MODE_APSTA;
     }
   else
-#endif /* ESP32_WLAN_HAS_STA */
+#endif /* ESPRESSIF_WLAN_HAS_STA */
     {
       mode = WIFI_MODE_AP;
     }
@@ -6006,7 +6006,7 @@ int esp_wifi_softap_stop(void)
 
   g_softap_started = false;
 
-#ifdef ESP32_WLAN_HAS_STA
+#ifdef ESPRESSIF_WLAN_HAS_STA
   if (g_sta_started)
     {
       ret = esp_wifi_set_mode(WIFI_MODE_STA);
@@ -6025,13 +6025,13 @@ int esp_wifi_softap_stop(void)
           goto errout;
         }
     }
-#endif /* ESP32_WLAN_HAS_STA */
+#endif /* ESPRESSIF_WLAN_HAS_STA */
 
   wlinfo("OK to stop Wi-Fi SoftAP\n");
 
-#ifdef ESP32_WLAN_HAS_STA
+#ifdef ESPRESSIF_WLAN_HAS_STA
 errout:
-#endif /* ESP32_WLAN_HAS_STA */
+#endif /* ESPRESSIF_WLAN_HAS_STA */
 
   esp_wifi_lock(false);
   return ret;
@@ -6670,7 +6670,7 @@ int esp_wifi_softap_rssi(struct iwreq *iwr, bool set)
   return -ENOSYS;
 }
 
-#endif /* ESP32_WLAN_HAS_SOFTAP */
+#endif /* ESPRESSIF_WLAN_HAS_SOFTAP */
 
 /****************************************************************************
  * Name: esp32_wifi_bt_coexist_init
@@ -6687,7 +6687,7 @@ int esp_wifi_softap_rssi(struct iwreq *iwr, bool set)
  *
  ****************************************************************************/
 
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
 int esp32_wifi_bt_coexist_init(void)
 {
   esp_coex_adapter_register(&g_coex_adapter_funcs);
@@ -6695,7 +6695,7 @@ int esp32_wifi_bt_coexist_init(void)
 
   return 0;
 }
-#endif /* CONFIG_ESP32_WIFI_BT_COEXIST */
+#endif /* CONFIG_ESPRESSIF_WIFI_BT_COEXIST */
 
 /****************************************************************************
  * Name: esp_wifi_stop_callback
