+++ arch/xtensa/src/esp32/esp32_wifi_adapter.h
@@ -45,20 +45,20 @@ extern "C"
  * Pre-processor Definitions
  ****************************************************************************/
 
-#if defined(CONFIG_ESP32_WIFI_STATION)
-#  define ESP32_WLAN_HAS_STA
-#  define ESP32_WLAN_STA_DEVNO    0
-#  define ESP32_WLAN_DEVS         1
-#elif defined(CONFIG_ESP32_WIFI_SOFTAP)
-#  define ESP32_WLAN_HAS_SOFTAP
-#  define ESP32_WLAN_SOFTAP_DEVNO 0
-#  define ESP32_WLAN_DEVS         1
-#elif defined(CONFIG_ESP32_WIFI_STATION_SOFTAP)
-#  define ESP32_WLAN_HAS_STA
-#  define ESP32_WLAN_HAS_SOFTAP
-#  define ESP32_WLAN_STA_DEVNO    0
-#  define ESP32_WLAN_SOFTAP_DEVNO 1
-#  define ESP32_WLAN_DEVS         2
+#if defined(CONFIG_ESPRESSIF_WIFI_STATION)
+#  define ESPRESSIF_WLAN_HAS_STA
+#  define ESPRESSIF_WLAN_STA_DEVNO    0
+#  define ESPRESSIF_WLAN_DEVS         1
+#elif defined(CONFIG_ESPRESSIF_WIFI_SOFTAP)
+#  define ESPRESSIF_WLAN_HAS_SOFTAP
+#  define ESPRESSIF_WLAN_SOFTAP_DEVNO 0
+#  define ESPRESSIF_WLAN_DEVS         1
+#elif defined(CONFIG_ESPRESSIF_WIFI_STATION_SOFTAP)
+#  define ESPRESSIF_WLAN_HAS_STA
+#  define ESPRESSIF_WLAN_HAS_SOFTAP
+#  define ESPRESSIF_WLAN_STA_DEVNO    0
+#  define ESPRESSIF_WLAN_SOFTAP_DEVNO 1
+#  define ESPRESSIF_WLAN_DEVS         2
 #endif
 
 #define SSID_MAX_LEN                (32)
@@ -160,7 +160,7 @@ void esp_wifi_free_eb(void *eb);
 
 int esp_wifi_notify_subscribe(pid_t pid, struct sigevent *event);
 
-#ifdef ESP32_WLAN_HAS_STA
+#ifdef ESPRESSIF_WLAN_HAS_STA
 
 /****************************************************************************
  * Name: esp_wifi_sta_start
@@ -496,9 +496,9 @@ int esp_wifi_sta_country(struct iwreq *iwr, bool set);
  ****************************************************************************/
 
 int esp_wifi_sta_rssi(struct iwreq *iwr, bool set);
-#endif /* ESP32_WLAN_HAS_STA */
+#endif /* ESPRESSIF_WLAN_HAS_STA */
 
-#ifdef ESP32_WLAN_HAS_SOFTAP
+#ifdef ESPRESSIF_WLAN_HAS_SOFTAP
 
 /****************************************************************************
  * Name: esp_wifi_softap_start
@@ -834,7 +834,7 @@ int esp_wifi_softap_country(struct iwreq *iwr, bool set);
  ****************************************************************************/
 
 int esp_wifi_softap_rssi(struct iwreq *iwr, bool set);
-#endif /* ESP32_WLAN_HAS_SOFTAP */
+#endif /* ESPRESSIF_WLAN_HAS_SOFTAP */
 
 /****************************************************************************
  * Name: esp32_wifi_bt_coexist_init
@@ -851,7 +851,7 @@ int esp_wifi_softap_rssi(struct iwreq *iwr, bool set);
  *
  ****************************************************************************/
 
-#ifdef CONFIG_ESP32_WIFI_BT_COEXIST
+#ifdef CONFIG_ESPRESSIF_WIFI_BT_COEXIST
 int esp32_wifi_bt_coexist_init(void);
 void coex_dbg_set_log_level(int level);
 #endif
