--- arch/xtensa/src/esp32/esp32_wifi_utils.h
@@ -1,104 +1,0 @@
-/****************************************************************************
- * arch/xtensa/src/esp32/esp32_wifi_utils.h
- *
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.  The
- * ASF licenses this file to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance with the
- * License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- ****************************************************************************/
-
-/****************************************************************************
- * Included Files
- ****************************************************************************/
-
-#ifndef __ARCH_XTENSA_SRC_ESP32_ESP32_WIFI_UTILS_H
-#define __ARCH_XTENSA_SRC_ESP32_ESP32_WIFI_UTILS_H
-
-#include <nuttx/config.h>
-#include <nuttx/net/netdev.h>
-
-#include <stdint.h>
-
-#ifndef __ASSEMBLY__
-
-#undef EXTERN
-#if defined(__cplusplus)
-#define EXTERN extern "C"
-extern "C"
-{
-#else
-#define EXTERN extern
-#endif
-
-/****************************************************************************
- * Public Function Prototypes
- ****************************************************************************/
-
-/****************************************************************************
- * Name: esp_wifi_start_scan
- *
- * Description:
- *   Scan all available APs.
- *
- * Input Parameters:
- *   iwr - The argument of the ioctl cmd
- *
- * Returned Value:
- *   OK on success (positive non-zero values are cmd-specific)
- *   Negated errno returned on failure.
- *
- ****************************************************************************/
-
-int esp_wifi_start_scan(struct iwreq *iwr);
-
-/****************************************************************************
- * Name: esp_wifi_get_scan_results
- *
- * Description:
- *   Get scan result
- *
- * Input Parameters:
- *   req      The argument of the ioctl cmd
- *
- * Returned Value:
- *   OK on success (positive non-zero values are cmd-specific)
- *   Negated errno returned on failure.
- *
- ****************************************************************************/
-
-int esp_wifi_get_scan_results(struct iwreq *iwr);
-
-/****************************************************************************
- * Name: esp_wifi_scan_event_parse
- *
- * Description:
- *   Parse scan information
- *
- * Input Parameters:
- *   None
- *
- * Returned Value:
- *     None
- *
- ****************************************************************************/
-
-void esp_wifi_scan_event_parse(void);
-
-#ifdef __cplusplus
-}
-#endif
-#undef EXTERN
-
-#endif /* __ASSEMBLY__ */
-#endif /* __ARCH_XTENSA_SRC_ESP32_ESP32_WIFI_UTILS_H */
